+++ julia_front/src/lib/config.js
@@ -28,4 +28,8 @@ export const API_ENDPOINTS = {
 	EXAM_PERIOD_BY_ID: (examPeriodId) => `${API_BASE_URL}/exam-periods/${examPeriodId}`,
 	// 관리자 메모 관련 엔드포인트
 	ADMIN_MEMO: `${API_BASE_URL}/admin/memo`,
+	// 과제 관련 엔드포인트
+	ASSIGNMENTS: `${API_BASE_URL}/assignments`,
+	ASSIGNMENTS_BY_USER: (userId) => `${API_BASE_URL}/assignments/user/${userId}`,
+	CLASS_ASSIGNMENTS: `${API_BASE_URL}/classes/assignments`,
 };
